# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Timeless Love API
  description: Backend API for Timeless Love family social platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@timelesslove.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.timelesslove.com
    description: Production server
...
paths:
...
  /api/v1/storage/upload-url:
    post:
      tags:
        - storage
      summary: Generate Upload URL
      description: Generate signed URL for uploading media to Supabase Storage
      operationId: generateUploadUrl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/storage/access-url:
    get:
      tags:
        - storage
      summary: Generate Access URL
      description: Generate signed URL for accessing media from Supabase Storage
      operationId: generateAccessUrl
      security:
        - bearerAuth: []
      parameters:
        - name: storage_path
          in: query
          required: true
          schema:
            type: string
        - name: expires_in
          in: query
          schema:
            type: integer
            default: 3600
      responses:
        '200':
          description: Access URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/storage/media/{media_id}/url:
    get:
      tags:
        - storage
      summary: Get Media URL
      description: Get signed URL for a specific media file by media ID
      operationId: getMediaUrl
      security:
        - bearerAuth: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: expires_in
          in: query
          schema:
            type: integer
            default: 3600
      responses:
        '200':
          description: Media URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessUrlResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadUrlRequest:
      type: object
      required:
        - memory_id
        - file_name
        - mime_type
      properties:
        memory_id:
          type: string
          format: uuid
        file_name:
          type: string
        mime_type:
          type: string
    UploadUrlResponse:
      type: object
      properties:
        upload_url:
          type: string
          format: uri
        storage_path:
          type: string
        expires_in:
          type: integer
    AccessUrlResponse:
      type: object
      properties:
        access_url:
          type: string
          format: uri
        expires_in:
          type: integer
    Error:
      type: object
      properties:
        detail:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...